Add conftest and move fixture there
